Redefine grammar-grammar in DSL
Static memory usage has increased by ~20kb, which I think is due to
extra class and function definitions.  The difference in memory before
and after defining the grammar is constant though (due to the fact that
the object in memory is identical to before).  It's much more succinct,
arguably more understandable, and way cooler, so I'll count it as a win.